+++ src/Tree/CacheEntry.java
@@ -5,13 +5,13 @@ public class CacheEntry {
     private int depth;
     private int nodes;
     private BinaryTree.Node optimalNode;
-    private double lowerBound;
+    private int lowerBound;
 
     public CacheEntry(int depth, int nodes) {
         this.depth = depth;
         this.nodes = nodes;
         this.optimalNode = BinaryTree.Node.createInfeasibleNode();
-        this.lowerBound = Double.MAX_VALUE;
+        this.lowerBound = Integer.MAX_VALUE;
     }
 
     public CacheEntry(int depth, int nodes, BinaryTree.Node node) {
@@ -28,15 +28,15 @@ public class CacheEntry {
         return optimalNode;
     }
 
-    public double getOptimalValue() {
+    public int getOptimalValue() {
         if (optimalNode != null && optimalNode.isFeasible()) {
             return optimalNode.mc;
         }
-        return Float.MAX_VALUE;
+        return Integer.MAX_VALUE;
     }
 
-    public void updateLowerBound(double lb) {
-        if (lb < Double.MAX_VALUE && lb >= 0 && ((lb <= lowerBound && optimalNode.isFeasible()) || !optimalNode.isFeasible())) {
+    public void updateLowerBound(int lb) {
+        if (lb < Integer.MAX_VALUE && lb >= 0 && ((lb <= lowerBound && optimalNode.isFeasible()) || !optimalNode.isFeasible())) {
             this.lowerBound = lb;
         }
     }
@@ -46,7 +46,7 @@ public class CacheEntry {
         this.lowerBound = node.mc;
     }
 
-    public double getLowerBound() {
+    public int getLowerBound() {
         return lowerBound;
     }
